hermes/routing/requirements.txt
# Requirements for Local-First Routing Module
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
pydantic>=2.0.0
